# Conflicts:
#	Forsikringsselskap/.idea/workspace.xml
#	Forsikringsselskap/target/classes/ARAprog/Controller.class

+++ Forsikringsselskap/src/main/java/ARAprog/Persons/Customer.java
@@ -9,7 +9,7 @@ public class Customer extends Person {
     int[] array;
 
     //Konstruktør for Customer
-    public Customer(int prsnPnr, String prsnAge, String prsnFirstName, String prsnLastName, String prsnAdress,
+    public Customer(long prsnPnr, String prsnAge, String prsnFirstName, String prsnLastName, String prsnAdress,
                     String prsnAreaCode, String prsnPostAlarea, String prsnEMail, String prsnTlfNr) {
 
         super(prsnPnr, prsnAge, prsnFirstName, prsnLastName, prsnAdress, prsnAreaCode, prsnPostAlarea, prsnEMail, prsnTlfNr);
